requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
Django==4.0.6
django-cors-headers==3.13.0
django-filter==22.1
django-rest-knox==4.2.0
django-restframework==0.0.1
django-six==1.0.4
djangorestframework==3.13.1
drf-yasg==1.21.3
et-xmlfile==1.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-client==0.0.1
mysqlclient @ file:///C:/Users/Admin/Downloads/mysqlclient-1.4.6-cp38-cp38-win32.whl
openpyxl==3.0.10
packaging==21.3
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
